requirements.txt
openai
pandas
pdfplumber
openpyxl
python-docx